+++ zarrs_data_type/src/data_type.rs
@@ -6,7 +6,7 @@ use derive_more::From;
 use half::{bf16, f16};
 use thiserror::Error;
 
-use crate::metadata::v3::array::{
+use zarrs_metadata::v3::array::{
     data_type::{DataTypeMetadataV3, DataTypeSize},
     fill_value::{
         bfloat16_to_fill_value, float16_to_fill_value, float32_to_fill_value,
@@ -356,7 +356,7 @@ impl core::fmt::Display for DataType {
 mod tests {
     use super::*;
 
-    use crate::metadata::v3::array::{
+    use zarrs_metadata::v3::array::{
         fill_value::{FillValueFloatStringNonFinite, HexString},
         nan_representations::{ZARR_NAN_BF16, ZARR_NAN_F16, ZARR_NAN_F32, ZARR_NAN_F64},
     };